Refactor LoanService to use LoanRequestDTO and improve business logic

+++ src/main/java/com/example/demo/entity/Loan.java
@@ -14,7 +14,7 @@ public class Loan {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "loan_id")
-    private Integer loanId;
+    private Long loanId;
 
     @ManyToOne(optional = false)
     @JoinColumn(name = "asset_id")
